removed utils out and added all the functionality to milvus_online_store
